Register a new app user
