+++ composer.json
@@ -13,11 +13,13 @@
             "App\\Entity\\": "src/Monofony/Bundle/CoreBundle/Recipe/src/Entity/",
             "App\\EventListener\\": "src/Monofony/Bundle/CoreBundle/Recipe/src/EventListener/",
             "App\\EventSubscriber\\": "src/Monofony/Bundle/CoreBundle/Recipe/src/EventSubscriber/",
-            "App\\Fixture\\": "src/Monofony/Plugin/FixturesPlugin/Recipe/src/Fixture/",
+            "App\\Fixture\\": "src/Monofony/Pack/FixturesPack/Recipe/src/Fixture/",
             "App\\Form\\": "src/Monofony/Bundle/CoreBundle/Recipe/src/Form/",
             "App\\Formatter\\": "src/Monofony/Bundle/CoreBundle/Recipe/src/Formatter/",
             "App\\Installer\\": "src/Monofony/Bundle/CoreBundle/Recipe/src/Installer/",
             "App\\Installer\\Checker\\": "src/Monofony/Bundle/CoreBundle/Recipe/src/Installer/Checker/",
+            "App\\Message\\": "src/Monofony/Pack/ApiPack/Recipe/src/Message/",
+            "App\\MessageHandler\\": "src/Monofony/Pack/ApiPack/Recipe/src/MessageHandler/",
             "App\\Migrations\\": "src/Monofony/Bundle/CoreBundle/Recipe/src/Migrations/",
             "App\\Repository\\": "src/Monofony/Bundle/CoreBundle/Recipe/src/Repository/",
             "App\\Validator\\": "src/Monofony/Bundle/CoreBundle/Recipe/src/Validator/",
@@ -48,7 +50,7 @@
             "App\\Tests\\Behat\\Exception\\": "src/Monofony/Bundle/CoreBundle/Recipe/tests/Behat/Exception/",
             "App\\Tests\\Behat\\Page\\Frontend\\": "src/Monofony/Bundle/FrontBundle/Recipe/tests/Behat/Page/Frontend/",
             "App\\Tests\\Behat\\Service\\": "src/Monofony/Bundle/CoreBundle/Recipe/tests/Behat/Service/",
-            "App\\Tests\\Controller\\": "src/Monofony/Bundle/ApiBundle/Recipe/tests/Controller/"
+            "App\\Tests\\Controller\\": "src/Monofony/Pack/ApiPack/Recipe/tests/Controller/"
         }
     },
     "require": {
