+++ config/packages/api_platform.yaml
@@ -1,7 +1,9 @@
 api_platform:
     version: !php/const App\Kernel::VERSION
     mapping:
-        paths: ['%kernel.project_dir%/src/Monofony/Bundle/CoreBundle/Recipe/src/Entity']
+        paths:
+            - '%kernel.project_dir%/src/Monofony/Bundle/CoreBundle/Recipe/src/Entity'
+            - '%kernel.project_dir%/src/Monofony/Pack/ApiPack/Recipe/config/api_platform/resources'
     patch_formats:
         json: ['application/merge-patch+json']
     swagger:
