--- config/packages/monofony_fixtures_plugin.yaml
@@ -1,2 +1,0 @@
-imports:
-    - { resource: "../../src/Monofony/Plugin/FixturesPlugin/Recipe/config/packages/monofony_fixtures_plugin.yaml" }
