--- src/Monofony/Bundle/ApiBundle/composer.json
@@ -1,22 +1,0 @@
-{
-    "name": "monofony/api-bundle",
-    "type": "symfony-bundle",
-    "description": "Monofony API bundle.",
-    "license": "MIT",
-    "require": {
-        "php": "^7.2",
-        "api-platform/api-pack": "^1.2",
-        "monofony/core-bundle": "^0.3"
-    },
-    "autoload": {
-        "psr-4": { "Monofony\\Bundle\\ApiBundle\\": "" },
-        "exclude-from-classmap": ["Recipe/"]
-    },
-    "minimum-stability": "dev",
-    "prefer-stable": true,
-    "extra": {
-        "branch-alias": {
-            "dev-master": "0.3-dev"
-        }
-    }
-}
