+++ src/Monofony/Bundle/CoreBundle/Recipe/src/Validator/Constraints/UniqueAppUserEmail.php
@@ -1,0 +1,25 @@
+<?php
+
+/*
+ * This file is part of mz_067_s_ccpa_thermotool.
+ *
+ * (c) Mobizel
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace App\Validator\Constraints;
+
+use Symfony\Component\Validator\Constraint;
+
+/**
+ * @Annotation
+ */
+final class UniqueAppUserEmail extends Constraint
+{
+    /** @var string */
+    public $message = 'sylius.user.email.unique';
+}
