+++ src/Monofony/Pack/ApiPack/Recipe/src/Message/RegisterAppUser.php
@@ -1,0 +1,77 @@
+<?php
+
+/*
+ * This file is part of mz_067_s_ccpa_thermotool.
+ *
+ * (c) Mobizel
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace App\Message;
+
+use Symfony\Component\Validator\Constraints as Assert;
+use App\Validator\Constraints as CustomConstraints;
+
+final class RegisterAppUser
+{
+    /**
+     * @var string
+     *
+     * @Assert\NotBlank(message="sylius.customer.first_name.not_blank")
+     */
+    public $firstName;
+
+    /**
+     * @var string
+     *
+     * @Assert\NotBlank(message="sylius.customer.last_name.not_blank")
+     */
+    public $lastName;
+
+    /**
+     * @var string
+     *
+     * @Assert\NotBlank(message="sylius.customer.email.not_blank")
+     * @Assert\Email(mode="strict", message="sylius.customer.email.invalid")
+     * @Assert\Length(
+     *     max=254,
+     *     maxMessage="sylius.customer.email.max"
+     * )
+     * @CustomConstraints\UniqueAppUserEmail()
+     */
+    public $email;
+
+    /**
+     * @var string
+     *
+     * @Assert\NotBlank(message="sylius.user.plainPassword.not_blank")
+     * @Assert\Length(
+     *     min=4,
+     *     minMessage="sylius.user.password.min",
+     *     max=254,
+     *     maxMessage="sylius.user.password.max"
+     * )
+     */
+    public $password;
+
+    /** @var string|null */
+    public $phoneNumber;
+
+    public function __construct(
+        string $firstName,
+        string $lastName,
+        string $email,
+        string $password,
+        ?string $phoneNumber = null
+    ) {
+        $this->firstName = $firstName;
+        $this->lastName = $lastName;
+        $this->email = $email;
+        $this->password = $password;
+        $this->phoneNumber = $phoneNumber;
+    }
+}
