+++ src/Monofony/Pack/ApiPack/Recipe/tests/Responses/Expected/authentication/new_access_token.json
