+++ src/Monofony/Pack/ApiPack/composer.json
@@ -1,0 +1,18 @@
+{
+    "name": "monofony/api-pack",
+    "type": "symfony-bundle",
+    "description": "A pack for Monofony.",
+    "license": "MIT",
+    "require": {
+        "api-platform/api-pack": "^1.2",
+        "monofony/core-bundle": "^0.3"
+    },
+    "autoload": {
+        "exclude-from-classmap": ["Recipe/"]
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-master": "0.3-dev"
+        }
+    }
+}
