+++ src/Monofony/Pack/FixturesPack/Recipe/src/Fixture/Factory/AbstractExampleFactory.php
@@ -1,0 +1,24 @@
+<?php
+
+/*
+ * This file is part of the Monofony package.
+ *
+ * (c) Monofony
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace App\Fixture\Factory;
+
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+/**
+ * Remove when Sylius fixtures bundle will include this.
+ */
+abstract class AbstractExampleFactory implements ExampleFactoryInterface
+{
+    abstract protected function configureOptions(OptionsResolver $resolver): void;
+}
