+++ src/Monofony/Pack/FixturesPack/Recipe/src/Fixture/Factory/ExampleFactoryInterface.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of the Monofony package.
+ *
+ * (c) Monofony
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace App\Fixture\Factory;
+
+/**
+ * Remove when Sylius fixtures bundle will include this.
+ */
+interface ExampleFactoryInterface
+{
+    public function create(array $options = []);
+}
