--- src/Monofony/Plugin/FixturesPlugin/MonofonyFixturesPlugin.php
@@ -1,20 +1,0 @@
-<?php
-
-/*
- * This file is part of the Monofony package.
- *
- * (c) Monofony
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Monofony\Plugin\FixturesPlugin;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class MonofonyFixturesPlugin extends Bundle
-{
-}
