--- src/Monofony/Plugin/FixturesPlugin/composer.json
@@ -1,22 +1,0 @@
-{
-    "name": "monofony/fixtures-plugin",
-    "type": "symfony-bundle",
-    "description": "Monofony fixtures plugin.",
-    "license": "MIT",
-    "require": {
-        "php": "^7.2",
-        "monofony/core-bundle": "^0.3",
-        "sylius/fixtures-bundle": "^1.4"
-    },
-    "autoload": {
-        "psr-4": { "Monofony\\Plugin\\FixturesPlugin\\": "" },
-        "exclude-from-classmap": ["Recipe/"]
-    },
-    "minimum-stability": "dev",
-    "prefer-stable": true,
-    "extra": {
-        "branch-alias": {
-            "dev-master": "0.3-dev"
-        }
-    }
-}
